+++ commuterFlows/utils/commuterOL.js
@@ -111,7 +111,8 @@ export class CommuterOL {
                     visibility: true,
                     type: "layer",
                     showLayerInTree: false,
-                    typ: "VECTORBASE"
+                    typ: "VECTORBASE", 
+                    alwaysOnTop: true
                 });
 
                 if (layer) {